It is the doc tests that are failing
I'm not entirely sure how to handle this.
It's fixable by just making the feature default.
However, there is this this: https://users.rust-lang.org/t/doctests-that-require-a-non-default-feature-is-it-possible/29529/5
Does that mean that the entirety of the observer docs should be wrapped like so?
#[cfg_attr(feature = "observers", doc = r##" ...docs..."##)]
Instead of /// ...

What I don't get about this, is that our observer docs are already within the observer mod tree which has been cfg'd away. I suppose that doesn't matter to the docs though.
Is just labelling it as a default feature the way to go then? It'll be default anyways, but I am unsure if that is how this should work.

I'm gonna take a look at the other bevy crates to see if I can find examples of how this is supposed to be done.