As default feature
This seems like the way to go.
Also typo for the usage statement (and I did check if that was the cause of the doc tests failing).

Now, we have a proof of concept of it being feature flagged. What is still up for consideration, the question of if this is clean enough.

In a perfect world you just cfg the observer mod tree, we do not live in that world. I will look around and see if I can make it better, but I think what we have here is good enough; I don't see it as very messy, but I'm not against the idea of making it more clean; I will see if there are any low hanging fruit.